# openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "example": "invalid_request",
                "type": "string"
              },
              "field": {
                "example": "startsAt",
                "type": "string"
              },
              "message": {
                "example": "startsAt must be a future date and time",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Payout": {
        "properties": {
          "amountMinor": {
            "example": 500000,
            "format": "int64",
            "type": "integer"
          },
          "eventId": {
            "example": "018f2c1e-9a11-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "hostId": {
            "example": "018f2c1e-1111-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "id": {
            "example": "018f2c1e-ffff-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "paidAt": {
            "example": "2026-12-01T19:01:00Z",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "providerPayoutId": {
            "example": "mock_pyt_018f2c1e-ffff-7c3b-9d4e-2b8a1c5f7e90",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "processing",
              "paid",
              "failed"
            ],
            "example": "paid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "hostId",
          "amountMinor",
          "status"
        ],
        "type": "object"
      },
      "PayoutPage": {
        "properties": {
          "nextCursor": {
            "example": "",
            "type": "string"
          },
          "payouts": {
            "items": {
              "$ref": "#/components/schemas/Payout"
            },
            "type": "array"
          }
        },
        "required": [
          "payouts",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "Refund": {
        "properties": {
          "amountMinor": {
            "example": 500000,
            "format": "int64",
            "type": "integer"
          },
          "eventId": {
            "example": "018f2c1e-9a11-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "id": {
            "example": "018f2c1e-eeee-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "providerRefundId": {
            "example": "mock_rfnd_018f2c1e-eeee-7c3b-9d4e-2b8a1c5f7e90",
            "nullable": true,
            "type": "string"
          },
          "purchaseId": {
            "example": "018f2c1e-aaaa-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "refundedAt": {
            "example": "2026-12-01T18:02:00Z",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "processing",
              "refunded",
              "failed"
            ],
            "example": "refunded",
            "type": "string"
          },
          "userId": {
            "example": "018f2c1e-bbbb-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "userId",
          "purchaseId",
          "amountMinor",
          "status"
        ],
        "type": "object"
      },
      "RefundPage": {
        "properties": {
          "nextCursor": {
            "example": "",
            "type": "string"
          },
          "refunds": {
            "items": {
              "$ref": "#/components/schemas/Refund"
            },
            "type": "array"
          }
        },
        "required": [
          "refunds",
          "nextCursor"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Opaque session token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend for paid live events: event lifecycle, ticket purchase, mocked payment, refunds, and payouts.",
    "title": "LIV DOT API",
    "version": "1.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/events/{id}/refund": {
      "post": {
        "operationId": "refundEvent",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "refunds": {
                      "example": 12,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "refunds"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Refund a failed event's viewers",
        "tags": [
          "finance"
        ],
        "x-roles": [
          "internal_admin"
        ]
      }
    },
...
    "/api/payouts": {
      "get": {
        "operationId": "listPayouts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutPage"
                }
              }
            },
            "description": "success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List payouts",
        "tags": [
          "finance"
        ],
        "x-roles": [
          "host",
          "internal_admin"
        ]
      }
    },
    "/api/payouts/{id}": {
      "get": {
        "operationId": "getPayout",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payout"
                }
              }
            },
            "description": "success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "not found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get one payout",
        "tags": [
          "finance"
        ],
        "x-roles": [
          "host",
          "internal_admin"
        ]
      }
    },
    "/api/refunds": {
      "get": {
        "operationId": "listRefunds",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundPage"
                }
              }
            },
            "description": "success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List refunds",
        "tags": [
          "finance"
        ],
        "x-roles": [
          "user",
          "internal_admin"
        ]
      }
    },
    "/api/refunds/{id}": {
      "get": {
        "operationId": "getRefund",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refund"
                }
              }
            },
            "description": "success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "not found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get one refund",
        "tags": [
          "finance"
        ],
        "x-roles": [
          "user",
          "internal_admin"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "local",
      "url": "http://localhost:8080"
    },
    {
      "description": "same origin",
      "url": "/"
    }
  ]
}